tests/wpt/web-platform-tests/app-history/currentchange-event/currentchange-location-api.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(t => {
  let oncurrentchange_called = false;
  let original_entry = appHistory.current;
  appHistory.oncurrentchange = t.step_func(e => {
    oncurrentchange_called = true;
    assert_equals(e.from, original_entry);
    assert_equals(e.from.index, -1);
    assert_equals(e.navigationType, "replace");
    assert_equals(appHistory.current.index, 0);
  });
  location.hash = "#foo";
  assert_true(oncurrentchange_called);
}, "AppHistoryCurrentChangeEvent fires for location API navigations");
</script>
